Subindo a API

+++ src/main/java/com/desafios/soccer/repository/JogadorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JogadorRepository extends JpaRepository<Jogador, Integer> {
+public interface JogadorRepository extends JpaRepository<Jogador, Long> {
 
 }
